fix: split TestGitChangedFilesLocal into test_test_filter_local_diff.py
test_test_filter.py exceeded the 1000-line arch constraint after adding
aggressive-mode tests. Move TestGitChangedFilesLocal to a dedicated file
to bring the original under threshold (988 lines).
